fix: revised approach to determining lo type
