+++ cli/tutors-gen/package.json
@@ -1,6 +1,6 @@
 {
   "name": "tutors-publish",
-  "version": "2.9.3",
+  "version": "2.9.4",
   "description": "Static course web site generator - json emitter",
   "main": "./src/tutors-gen.js",
   "scripts": {
